Alteracoes para apenas ROS1

+++ src/lds_lidar.cpp
@@ -187,7 +187,7 @@ void LdsLidar::SetLidarPubHandle() {
   pub_handler().SetPointCloudsCallback(LidarCommonCallback::OnLidarPointClounCb, g_lds_ldiar);
   pub_handler().SetImuDataCallback(LidarCommonCallback::LidarImuDataCallback, g_lds_ldiar);
 
-  double publish_freq = Lds::GetLdsFrequency();
+  uint8_t publish_freq = Lds::GetLdsFrequency();
   pub_handler().SetPointCloudConfig(publish_freq);
 }
